messager: consolidate end2end tests into messaging
